ZOOKEEPER-3430: Observability improvement: provide top N read / write path queries.
We would like to have a better understanding of the type of workloads hit ZK, and one aspect of such understanding is to be able to answer queries of top N read and top N write request path. Knowing the hot request paths will allow us better optimize for such workloads, for example, enabling path specific caching, or change the path structure (e.g. break a long path to hierarchical paths).

This commit adds a RequestPathMetricsCollector which will collect path requests stats and answers such queries. There are also a set of new Java system properties that make this collector highly configurable, including disable it totally (the default option is on.).

The stats can also be queried through commands, which I will add in a different pull request once this landed, since this one is already big.

Author: Michael Han <lhan@twitter.com>

Reviewers: Enrico Olivelli <eolivelli@apache.org>

Closes #989 from hanm/twitter/065e164bfbbf467a28602da76e9443faacaec1c8