ZOOKEEPER-3124: Add the correct comment to show why we need the special logic to handle cversion and pzxid
There is special logic in the DataTree.processTxn to handle the NODEEXISTS when createNode, which is used to handle the cversion and pzxid not being updated due to fuzzy snapshot:

https://github.com/apache/zookeeper/blob/master/src/java/main/org/apache/zookeeper/server/DataTree.java#L962-L994.

But seems this is not a real issue, in the current code, when serializing a parent node, we'll lock on it, and take a children snapshot at that time. If the child added after the parent is serialized to disk, then it won't be written out, so we shouldn't hit the issue where the child is in the snapshot but parent cversion and pzxid is not changed.

But maybe I'm missing something, there is not much discussion in the Jira, so create a PR to have more attention.

Author: Fangmin Lyu <allenlyu@fb.com>

Reviewers: Norbert Kalmar <nkalmar@apache.org>

Closes #610 from lvfangmin/ZOOKEEPER-3124