ZOOKEEPER-1936: Server exits when unable to create data directory due to race
Resurrecting an ancient ticket which could be fixed with a simple patch.

Jira mentions a scenario when auto purging tool is in use and Zookeeper server could have a race condition when creating snapshot and data directories. (directory auto creating is enabled by default)

Double checking the directory existence might help with it.

Author: Andor Molnar <andor@apache.org>

Reviewers: Enrico Olivelli <eolivelli@apache.org>

Closes #1225 from anmolnar/ZOOKEEPER-1936